priorityRX-agent-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
celery[redis]
redis
httpx
sqlalchemy
python-dotenv
mlflow
boto3
joblib
scikit-learn
pandas
numpy